Use `floor` instead of `round`
WebAssembly's `f32.nearest` function doesn't behave exactly the same as Rust's `f32::round`, which results in a slower version getting used instead. So, switch to `f32::floor` which does take advantage of the native `f32.floor` instruction.

The benchmarks claim that this is also faster on native but it also claims that the non-wrapping code is faster too which I didn't touch, so I don't trust it.